feat(frontend): add mobile adaptation for /devices and /devices/chat pages
- Split /devices and /devices/chat pages into Desktop/Mobile components
- Add device selector to mobile sidebar for /devices/chat page
- Add Beta badge to mobile top navigation for devices pages
- Create useDeviceStatusHelpers hook for shared device status logic
- Extend TaskSidebar with device selection props for mobile mode
- Touch-friendly controls with 44px minimum targets
- Simplified device cards for mobile with name and status only
